MSI: Move SPI interfaces for export to model directory
